*: refine depency of contexts
