+++ br/pkg/lightning/backend/kv/BUILD.bazel
@@ -24,6 +24,7 @@ go_library(
         "//br/pkg/utils",
         "//pkg/errctx",
         "//pkg/expression",
+        "//pkg/infoschema/context",
         "//pkg/kv",
         "//pkg/meta/autoid",
         "//pkg/parser/model",
