+++ pkg/expression/BUILD.bazel
@@ -74,6 +74,7 @@ go_library(
         "//pkg/config",
         "//pkg/errctx",
         "//pkg/errno",
+        "//pkg/expression/context",
         "//pkg/extension",
         "//pkg/kv",
         "//pkg/parser",
@@ -86,7 +87,6 @@ go_library(
         "//pkg/parser/terror",
         "//pkg/parser/types",
         "//pkg/privilege",
-        "//pkg/sessionctx",
         "//pkg/sessionctx/stmtctx",
         "//pkg/sessionctx/variable",
         "//pkg/types",
