+++ pkg/expression/context/BUILD.bazel
@@ -3,13 +3,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "context",
     srcs = ["context.go"],
-    importpath = "github.com/pingcap/tidb/pkg/planner/context",
+    importpath = "github.com/pingcap/tidb/pkg/expression/context",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/expression/context",
+        "//pkg/infoschema/context",
         "//pkg/kv",
-        "//pkg/parser/model",
         "//pkg/sessionctx/variable",
-        "//pkg/util/context",
     ],
 )
