+++ pkg/infoschema/context/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "context",
+    srcs = ["infoschema.go"],
+    importpath = "github.com/pingcap/tidb/pkg/infoschema/context",
+    visibility = ["//visibility:public"],
+)
