+++ pkg/session/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//pkg/extension",
         "//pkg/extension/extensionimpl",
         "//pkg/infoschema",
+        "//pkg/infoschema/context",
         "//pkg/kv",
         "//pkg/meta",
         "//pkg/metrics",
