+++ pkg/table/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "//pkg/errctx",
         "//pkg/errno",
         "//pkg/expression",
+        "//pkg/infoschema/context",
         "//pkg/kv",
         "//pkg/meta/autoid",
         "//pkg/parser",
